ci(spotbugs): add a SpotBugs exclude filter to avoid failing builds.
The bugs EI_EXPOSE_REP and EI_EXPOSE_REP2 are filtered now. These are now
following a positive-list pattern (instead of a negative-list before v4.3.0)
and nag on potential mangling on objects from untrusted code.

As users of this lib are to implement the interface classes, it does not make
sense to speak about "untrusted" here...